fix: 컴포넌트 분리 리팩토링
